Audit Plugin Modal Fix
This commit removes a line of code that caused the Ticket Audit modal to close any time you tried to go to the next page of audits or sort the audits by timestamp. The code was put in place 6 years ago for file upload fields, but they still work fine everywhere in the codebase without that line now.