issue: Mass Process Add Users To Organization
This addresses an issue where checking the box next to a few User’s names
then selecting the Action to Add To Organization does nothing. This is due
to a pesky jQuery issue where `parseJSON()` will crap out if you pass in
JSON (go figure). You MUST provide a string otherwise it will silently fail.
So, this adds an if statement that checks if the content is a string and if
so we will use `$.parseJSON()` otherwise we will not parse it.