issue: Confirm Popup Promise
This addresses an issue where clicking Cancel on some popups (eg.
Organization Deletion popup, User Deletion popup, etc.) will perform the
action instead of cancelling. This is because the `then()` and `done()`
callbacks used on the custom `$.confirm()` method do not respect the
returned promise. This adds a check to each `$.confirm()` usage to see if
the promise is equal to `FALSE`, if so we return false. If the promise is
not equal to `FALSE` then we continue with the callback.