Issue: New Agent Extended Access and Teams
This commit fixes an issue we had with trying to create new Agents with Extended Department Access or a Team. We were trying to add them before there was a staff_id which resulted in a DB error as well as the actions not being performed. To fix this, we just need to add the Agent's Department access and Team membership after we save the Agent so that they have an id. We also need to save the Staff object in the setDepartmentId function.