feature: Configurable Agents As Collaborators
This adds a new setting called "Disable Agent Collaborators" that allows Admins
to globally configure wether or not to allow Agents to be added as
Collaborators via any communication on tickets. In some use cases Agents do
not want to be added as Collaborators on tickets to avoid an overload of
emails/alerts, etc. This adds a new checkbox labeled `Disable Agent
Collaborators` to **Admin Panel > Settings > Agents** that if
Enabled/Checked will automatically Disable Agents that are added as
Collaborators. This disables the participant alerts for the Agents whilst
giving other Agents/Admins the ability to enable them at a later date
instead of having to manually add them back. If an Agent tries adding
another Agent as a Collaborator via the Manage Collaborators modal it will
also Disable the Agent Collaborator.