collab: Refuse To Add Owner As Collaborator
This adds a new `if()` statement to `Thread::addCollaborator()` that checks
the current Ticket Owner's ID to the User's ID that is being added as a
Collaborator. If the IDs match then we will refuse to add the Owner as a
Collaborator as Users are not supposed to be a Collaborator on Tickets they
own.