issue: Edit User Popup Perm
This is a rewrite of 5256 that addresses an issue where if an Agent does not
have permission to Edit Users the Edit User link on Ticket Create will show
a broken/blank popup. This adds a check for the Agent's User Edit Permission
and if they do not have permission to Edit Users we will remove the
`onclick` event so it does not trigger a popup.