xss: Response & Internal Note
This addresses an issue found when investigating a separate vulnerability
where XSS can be executed in the Response and Internal Note Fields when
errors are returned or if they are rendered with data (like drafts). This
adds sanitization to both Response and Internal Note fields for the New
Ticket page and the Ticket View page. This will sanitize both POST data and
Drafts to ensure all content is safe before rendering. This also adds
sanitization to the ThreadEntryWidget so any ThreadEntryField is safe from
XSS upon rendering (like the New Ticket Issue Summary in the Client Portal).