xss: List Values
This addresses an issue reported by Security Innovation where the "Plural
Name" value is not sanitized which results in XSS possibilities. Upon
further investigation, none of the Create List values are sanitized so any
of them could potentially house XSS. This adds sanitization to `update()`,
`create()`, and `add()` so that all content is safe from XSS before reaching
the database and/or being returned in errors. In addition, this adds a new
argument to `create()` called `$sanitize` that determines wether or not to
sanitize the content. This argument is `TRUE` by default. Lastly, this
sanitizes the `$_POST` vars when updating the List Properties.