upgrade: Redactor 3.3.5
This upgrades Redactor from `3.2.2` to `3.3.5` so that we are utilizing the
latest code. This addresses a few security issues with upstream Redactor as
well as provides a few bug fixes and improvements. This also updates
Redactor's core plugins as well as the CSS to ensure everything Redactor is
fully up-to-date.