issue: Collab Pass By Reference
This addresses an issue introduced with 5f5403d where passing `$vars` by
reference to `Thread::addCollaborator()` broke a couple of things for
Collaborators. This reverts the pass by reference and adds a new global
variable to class Collaborator called `active`. This allows us to track the
`active` variable whilst not messing with other collab functions.