revert: Agent Added Agent Collabs
This adds a check for `$thisstaff` in `Thread::addCollaborator()`. If we
have a staff object we will enable Agent Collabs, if we do not have a staff
object we will disable Agent Collabs. This will ensure all Agent added Agent
Collabs are Enabled and any User added Agent Collabs are Disabled. This also
updates the Help Tip for the setting to remove the `Web created tickets`
part (as users cannot add collabs via web) and to clarify that this only
affects User created tickets.