issue: AuditEntry Not Found
This addresses an issue where not having the Audit Log plugin installed and
attempting to delete a Help Topic, Knowledgebase Category, or Page the
system fails with the error `Class 'AuditEntry' not found`. This adds a
check for the `AuditEntry` class before calling it (for Help Topics,
Knowledgebase Categories, and Pages) to avoid the `Class not found` error.