issue: Move Owner Check to Ticket Collabs
This addresses an issue where adding a Task Collab fails with error `Call to
undefined method Task::getOwnerId()`. This updates
`ThreadAjaxAPI::addCollaborator()` to call `$object->addCollaborator()`
instead of `$thread->addCollaborator()`. This gives us the opportunity to
perform separate actions/checks for Ticket and Task Collabs. This also
updates `Ticket::addCollaborator()` to check if the current Collab is the
Owner, if so we will refuse to add the Owner as a Collab.