issue: Show Custom Validation Message
This addresses an issue where custom validation messages are not appearing,
instead the system shows the default validation messages. This is due to
9e21dfd where we failed to consider custom validation messages. Basically,
the referenced commit made it to where we could use validation messages from
the source validation method which is great but it overwrites the custom
validation message. The end goal is to have Custom Messages displayed first,
if no custom messages then we display the Source Messages, and if no source
messages we display the Default Messages. This renames the Default
Validation Message variable from `$error` to `$ferror` (for Function Error)
and updates the ternary check to first check `$error` (Custom Message) then
check `$err` (Source Message) then lastly `$ferror` (Default Message).