issue: Mass Delete Help Topics Warning
This addresses an issue where deleting 2 or more Help Topics shows a warning
of `X of X selected help topics deleted` even though all selected Help
Topics were deleted. This is due to the check that ensures that the number
of Help Topics deleted is the same as the how many Help Topics you selected.
This updates the check to compare against the actual Topic count rather than
the QuerySet itself.