issue: Countable Warning
This addresses an issue where using PHP 7.4, setting Help Topic Sort to
Manually, rearranging the Help Topics, and Saving Changes throws a warning
of `Parameter must be an array or an object that implements Countable`. This
is because when sorting Help Topics, we send no IDs so the `count()`
obviously fails. This checks to see if `$_POST['ids']` exists first, if so
then we count, otherwise we default to `0`.