xss: FAQ Category On Errors
This addresses 1/2 of the vulnerability reported at 5514 where pasting an
XSS Payload in the Category Description/Notes and saving changes without
filling out required info executes the XSS Payload. This is due to the
system not sanitizing the return values on errors. This adds
`Format::sanitize()` to the Description and Notes fields so that any value
displayed is properly sanitized. The Description/Notes are properly
sanitized before saving to the database, the issue only lies when returning
values on errors.