issue: Filter Action Add Button
This addresses issue 5513 where clicking the Add button on the Filter
Actions tab without selecting an option will insert an `undefined` Action.
This is due to the placeholder option having an empty `value` attribute.
This updates the JS to check the selected option's `value` attribute and if
empty we will return without appending. This will ensure the placeholder
option does not append an `undefined` Action.