print: Update Icons and Add Titles
This updates the icons for the Print options to be more explanatory of what
you will receive. PDF Files will have File Icons and ZIP Archives will have
Folder Icons. In addition this adds `title` attributes to all Print options
to note if the Agent will receive a PDF File or ZIP Archive.