templates: %{ticket.thread.complete} Not Respecting Identity Masking
This addresses an issue where Karen is not respecting the drip…just kidding.
This addresses an issue where `%{ticket.thread.complete}` does not respect
Agent Identity Masking. This adds a check for `hideStaffName()` and if
`TRUE` we will use `Staff` as the Agent's name in the thread export template
for clients. If `hideStaffName()` returns `FALSE` then we will use the
Agent's actual name. This only applies to the `thread-export.tmpl.php` for
clients.