issue: Signature Box No Longer Expands
This addresses issue 5531 where the Signature Box no longer expands when
hovering over it. This is due to some outdated jQuery that needs to be
updated. This updates `.on('hover'` and breaks it into `.on('mouseenter'`
and `.on('mouseleave'`. This also removes `originalHeight = Math.max` as it
just kept adding more and more pixels causing the collapsed area to grow and
grow (ew). In addition, this adds 25 pixels to the offsetHeight to expand
the box past the content and subtracts 10 pixels from originalHeight so it
collapses back to the same height.