sla: Force Intval For Scientific Floats
This addresses issue 5532 where some floats are being formatted in
Scientific Format (ie. `2.777E+4`) which causes a fatal error for
`DateInterval::__construct()`. This is most likely due to the `precision`
INI setting in some instances. This adds `intval()` around `round()` to
ensure an INT value is returned to avoid the fatal error on format issues.