inline: RichText Fields View First
This addresses an issue where things such as really long tables in inline
RichText Fields causes the remainder of the page to be malformed. This is
due to the content being truncated causing the broken HTML to interfere with
the page. This adds a check for RichText Fields when displaying inline and
if one, strips all tags. With the content being stripped, this adds a View
Popup so Agents can view the entire, true content. From the View Popup,
Agents can click the Edit button to edit the content. This adds a new Field
View template as well as AJAX regex/methods for returning the View Popup.