events: LongAnswer Fields
This addresses an issue where some LongAnswer fields with really long
content cut off the thread event `data`. This is due to the content being
longer than the max allowed characters for the `data` column. This adds
`Format::striptags()` and `Format::truncate()` to the TextareaField changes
to ensure they have no HTML and to ensure they are below the max allowed
characters. Stripping tags and truncating is done after we save the field
changes as if we do it beforehand, it will save that as the changes instead
of the actual changes.