i18n: Redactor Files Not Included
This addresses an issue where upon building the language packs, the Redactor
i18n files are not included. This is due to the wrong file paths being used
to check if the file exists and to get the file contents. This updates all
the file paths to point to the correct directory.