issue: Get Team Members For Alerts
This addresses an issue where when sending Alerts to Team Members, we grab
all Members regardless of if they have the Alerts checkbox Enabled or not.
This adds a new method called `getMembersForAlerts()` that will retrieve a
list of all Team Members that have the Alerts checkbox Enabled. This updates
all of the `getMembers()` calls to `getMembersForAlerts()` when sending
Alerts.