issue: Track Fields Disabled By System Default Help Topic
This addresses an issue where Tickets that get assigned the System Default
Help Topic (mainly emails) have Disabled Fields (disabled by Help Topic)
appear in the Client Portal on Edit. This is due to not running through the
Disabling of Fields for the Default Help Topic. This moves the section to
track Disabled Fields to a new method called `trackDisabledFields()` in
class Topic that can be reused. This also adds a check for the Default Help
Topic and if set, we will run through the new method.