visibility: Fields Disabled By Help Topic Appear In Ticket Header (Client Portal)
This addresses an issue where having a Field that is Disabled by Help Topic
and has a Default Value appears in the Ticket Header in the Client Portal.
This is due to the system not checking for Disabled Fields per Help Topic,
rather just checking for any entries with values. This adds a check to see
if any Fields are Disabled by Help Topic, if so, we will exclude them from
the Ticket Header in the Client Portal. This adds a new method called
`getDisabledFieldsIds()` that retrieves an array of Form ID and disabled
Field IDs (eg. `array(form_id => array(disabled_field_ids))`).