oops: Task Missing Parentheses
This addresses an issue where there is a missing parentheses in an upgrade
task. This is due to me 3heading the commit and forgetting to run the lint
tests so the missing parentheses was not revealed. This adds the missing
parentheses so the lint tests pass and there are no PHP errors/warnings.