issue: Agent Password Reset With No Existing Password
This addresses an issue where if an Agent never finishes registering their
account via the link in the Registration Email and tries to reset their
password, the system says `Unable to reset password. Contact your
administrator`. This is due to no password being set in the first place; we
cannot reset a password that doesn't exist. This adds a check for the
Agent's password when requesting password reset and if none set we will
resend the Registration Email. This will ensure the Agent can set a password
and login successfully.