cli: Manage.php Errors
This addresses three issues, the first of which is 5629 where running `$ php
manage.php` throws the error `Cannot declare class Exporter, because the
name is already in use`. The second issue is when addressing the declare
error you receive the error `Uncaught Error: Call to a member function
query() on null`. The last issue is when running `$ php manage.php export`
we receive the warnings `GROUP_TABLE: Cannot export table with no fields`
and `ost_timezone: Cannot export table with no fields`. This addresses the
first issue by renaming the Exporter class used in the CLI from `Exporter`
to `CliExporter` and updating all references to the class. This addresses
the second issue by using `Bootstrap::connect();` to ensure we have a
database connection before calling the Exporter. This addresses the last
issue by removing the references to `GROUP_TABLE` and `TIMEZONE_TABLE` in
the DatabaseExporter class. These tables are not used any longer so it’s
okay to remove the references.