issue: Custom REGEX Failure
This addresses an issue reported on the Forum where when using a Custom
REGEX Validator like `/^[0]\d{6}$/iu` and someone inserts a value like
`0000000` the REGEX will fail. This is due to a check where we use the
equals operator (`==`) to check if the value is equal to `0` and if so we
change the value to `&#48`. This is a problem for values such as `0000000`
as technically `0000000` is equal to `0`. This updates the equals comparison
operator to the identical comparison operator to ensure that only values
that are identical to `0 (string)` are changed to `&#48`.