Issue: Ticket Export Delimiters
This commit fixes an issue we had with changing the delimiter when exporting Tickets as a CSV. In order to save the delimiter to the Agent's session, we needed to match the input field name of 'csv-delimiter' rather than just 'delimiter'. This just makes it so that if the Agent has to do another export, they won't have to reset the delimiter again. Also, in order to correctly pass the delimiter to the options of the CsvExporter class, the delimiter needed to be added to the options array before constructing the new exporter.