Issue: Canned Response, Marked Answered
This commit fixes an issue where if an Agent creates a Ticket on behalf of a user and includes a Canned Reply, the Ticket is opened as already being answered. The postReply method is supposed to return the response and then go back to the postCannedReply function, and then the next line would mark it as unanswered, but it was never making it back into postCannedReply.