Issue: Prevent Deleting All Topics
This commit makes sure that we always have at least 1 Public and Active Help Topic

Mass Process:
- Do not allow deletion of all Topics
- Do not allow archiving/disabling all Topics
- Allow changing Topics from Disabled to Archived and vice versa

Note: If you try to archive, disable, or delete all Topics OR all Active Topics at once, it will throw the error that at least 1 topic must be active.

Topic Update:
- Do not allow changing Topic to Private if it's the only one that's Public
- Do not allow archiving/disabling if it's the only active Topic
- Allow changing Topics from Disabled to Archived and vice versa

It also gets rid of code we had in place in the Client Portal that would put a 'General Inquiry' Topic in the list if all Topics had been deleted. It didn't have an id to get a Topic form, so it still wouldn't allow for Ticket creation.