redactor: Improper Formatting When Double Spacing
This addresses an issue where if using Double Spacing in Redactor and you
put 2 or more lines in between paragraphs, the final content shows single
spacing formatting. This is due to Redactor using `<p></p>` for the double
spaced new lines and HTMLawed does not like that. This adds a new Format
function called `editor_spacing()` that uses simple regex to find and
replace `<p></p>` with `<p></br></p>` to ensure the formatting stays as
typed.