Authcode: Ticket Access Link
osTicket provides ability to add ticket access link that can auto-login
users (if auth codes feature is enable), but the auth code is not added
when a ticket has collaborators. This is on purpose to avoid access
leak, because one email is sent to all recipients. As workaround, a user
can request ticket Access Link, individually, but the system still
assumed that multiple recipients are getting the link based on number of
collaborators on the ticket - meaning auth code is not added.

This commit addresses the issue by making sure Access Link requests
bypass collaborators check.