issue: Draft Save
This addresses a small issue introduced with 5684 where Drafts do not save
when creating a New Ticket via Staff/Client side. This is due to a field
name mismatch in `include/ajax.draft.php`. We are expecting `message` but we
get the hashed name instead. This addresses another issue where Drafts do
not save when creating a New Task. We are expecting `description` but we get
the hashed name instead. This adds the hashed names to the array of possible
fields which allows the system to find the Draft bodies and save the Drafts
successfully.