security: Parent Ticket Access (Client)
This addresses a vulnerability reported by Daniel Treffenstädt (MuonPi Team)
where any User can get access to any Parent Ticket by ID (not Number) even
if they don't have access to the Parent or any Children. This is due to
checking if the ticket is a parent and is merged (not linked) without
checking for access to the children. This updates the check to grab all
children and check the User's access to each. If the User has access to one
of the children they are allowed access to the Parent, however if they do
not have access to one of the children they are denied access.