feature: Client Portal Open/Answered Queue
This feature adds a new Queue to the Client Portal called `Answered`. The
idea is the exact opposite of the Answered Queue for Agents, meaning,
"Answered" on the Client Portal are Tickets answered by the User (not
Agent). This helps further separate Tickets that need immediate attention or
a response from a User vs Tickets that have already been responded to by a
User. This Queue acts the same as Open and Closed in regards to if there are
no Answered tickets the Queue is hidden.