issue: Client-Side Reply Draft Saving
This addresses an issue where Draft saving for the Post a Reply box on the
Client Portal returns the error `Draft body not found in request`. This is
due to using the generic name of `message` instead of the hashed Form name.
This updates the code to set the hashed Form name using
`$messageField->getFormName()`.