issue: SCP Login Redirect
This addresses issue 5741 where when logged in as an Agent and visiting
`scp/login.php` directly it does not redirect you to the Ticket Queues,
instead it shows the login modal. This adds a check to `scp/login.php` to
see if there is a valid Agent Session and if so we redirect to the Ticket
Queues.