Issue: Echoing Default Dept Status
This commit fixes an issue introduced with PR #5755 where we just echo the status of the Default Department so that you don't have the option to disable/archive it. When we update a Department, if $vars['flags'] isn't present, we clear out the flags to 0 and then set them according to the other vars set. Since we are just echoing the status name for the default department, $vars['status'] is NULL, so when you save the department, it ends up being disabled (we can't put the flag back). So if we're updating the default department, we need to confirm that it is the default and then manually set $vars['status'] to 'active' so that the flag is set correctly.