issue: Task last_update Var
This addresses a small issue reported on the Forum where when using the
`%{task.last_update}` variable in the Task Email Templates you receive a
fatal error of `Uncaught OrmException: Task: last_update: Field not defined
in`. This is due to the variable pulling data from `last_udpate` whilst not
having a column called `last_update` in the `ost_task` table. This updates
the variable to pull data from `updated` as this is the correct column to
use in the `ost_task` table.