issue: getDefaultDeptId() On Null
This addresses an issue where the system throws a fatal error of `Fatal
error: Uncaught Error: Call to a member function getDefaultDeptId() on null`
when doing a fresh installation. This was introduced with 5777 and is due to
trying get the default Department's ID from the system config when it's not
available. This adds a check for `$cfg` before calling it to ensure no fatal
errors are thrown.