Issue: Agents/Depts in Queue Cols
This commit fixes an issue with how Departments and Agents show up in queue columns. If an Agent's Visibility was set to where they could only see Depts/Agents in a Department they have access to, it became an issue if the Agent was assigned a ticket in a Department they couldn't access (queue column would be blank) or if a Ticket was assigned to an Agent's Department but was assigned to an Agent that is not in any of the Agent's Departments (queue column would be the staff_id).

As long as an Agent has access to a Ticket, they should be able to see the value of the Dept/Agents related to that Ticket.