Issue: Visibility Permissions
This commit fixes an issue with checking on choices for the DepartmentChoiceField. We were setting $this->_choices to an empty array in the getQuickFilterChoices function, so when we did the check to see if choices were set or not, they were technically already set which resulted in the quick filters for the department field always showing as none.