Issue: Task Inline Transfer
This commit fixes an issue where we were not checking role permissions for inline edit on Tasks. If an Agent without the role permission to Transfer Tasks tried to click the Deparment to transfer it, it popped up a blank modal and said 'Permission Denied'. Now, the Department will not have the edit link if they don't have the permission.