issue: markAs Popup Manager (No Access)
This addresses an issue where if an Agent does not have access to a
Department but is the Manager of the Department and they go to mark a Ticket
as Answered/Unanswered a blank popup appears. This is due to `isManager()`
only checking the Agent’s Primary Department for Manager access. This
updates the method to accept a Department object and if one is passed we
will check this for Manager access, otherwise we will default to the Agent’s
Primary Department.