xss: onerror Property
This mitigates a vulnerability reported by Yunus AYDIN where some field
types are vulnerable to XSS via the `onerror` property. This adds `on*` to
the HtmLawed config to remove all of the `on*` properties from content. This
also updates `Format::htmlchars()` in some places to force sanitization of
the content.