Issue: Spaces in List Items
This commit fixes an issue we had with list items that began or ended with a space. If the list was being used in a Typeahead or Input field, the values couldn't be selected if they didn't match with the spaces exactly. This commit strips the leading and trailing spaces before saving the list items.