Issue: Custom Dept/Custom Assignee Exports
This commit fixes an issue with exporting Custom Department or Custom Assignee fields. In the exports, the fields were exporting as IDs instead of the actual field value.

Custom Department Field:
In from_query, if the value is just a dept_id instead of the department name, the field should pass the display function.

Custom Assignee Field:
- In to_php, we were missing logic needed to parse cdata into an array if the value was saved as 'id,value'
Ex: t34,Team1
First, we need to make sure not to accidentally clear out this value if the JsonDataParser doesn't work.
Next, this string needs to be set to an array like this in order to render properly:
['t34'] => 'Team1'

- If there is an Assigntee type, we should use substr to remove the letter from the array key in the event that the id for the field is more than one number
Ex: t34
The type should be 't' and the id should be 34.
Using key($value)[1] would set the id to just 3

- If the value coming into AssigneeField::to_php is the staff_id, meaning the value is just a number and we are not sending a type through, we should look up and return the Staff object.